Add a public domain dedication
The "software license" doesn't actually give you a license to do stuff with the software so far. Adding one line that dedicates it to the public domain.

The reason I say "Unless specified otherwise" is that I don't want to prevent someone from adding, say, MIT- or BSD-licensed code to the project.