Automatically reset vote counts when a PR is modified.
"Modifications" could be due to sneakiness, but could just be the
sensible way of resolving merge conflicts (e.g., you could notice
a merge conflict and rebase on top of `master`). This shouldn't
cause the whole PR to be wasted; it should simply reset the votes.

Only those votes posted since the most recent SHA-hash blessed by
the bot will be counted.  A SHA-hash is "blessed" if the bot has
posted it as the last word of a comment. (And only the bot; no other
users' comments can bless hashes.)

Notice that I'm removing a couple of unnecessary callback-style
interfaces. I'd eventually like to remove all of the callbacks
and replace them with straight-line code and/or event handlers;
but I refuse to do that until the
"merge conflict -> rebase -> PR auto-closed" pipeline is fixed.
Resetting votes instead of auto-closing PRs is the first step
in that direction.